DEV: Add "ruff check" to the pre-commit hooks.
This commit ensures "ruff check" is run on any files to be committed, so
that we don't end up with linting errors in CI, or a bunch of tiny
commits fixing style errors. It does *not* run "ruff check --fix" or
"ruff format", so no files will be changed unless done so explicitly by
the developer. Also adds "pre-commit" as a dev dependency.