ENH: Implement the error_handler function.
This function is largely the same as that used in the original
cholmod.pyx, but is now safe for use with a `noexcept` C interface,
since we just check for and handle the errors at the python level.

The "#cython: legacy_implicit_noexcept = True" line is not recommended
in modern code, so using `cholmod_common.error_handler` is not possible
(and not necessary) in our updated interface.

Add more informative error messages.