ENH: Use cholmod_resymbol_noperm to take permuted input matrix.
This update makes the CholeskyFactor.resymbol method API identical to
the procedural ldlupdate interface. It is likely more consistent with
how the function will actually be used, where the user has already
permuted the input matrix with the fill-reducing permutation by the time
multiple up/downdates have been computed and a resymbol is needed.