MNT/TST: Remove CholeskyFactor._convert_factor method.
This commit removes the unnecessary _convert_factor method. It does not
need to be called prior to cholmod_update, as that function convert the
factor internally if necessary. The conversion LL <-> LDL flag has been
moved to the _csc_view_from_cholmod_factor function, and incorporated
into the get_factor method.