FIX/TST: Force all input matrices to be canonical CSC format.
This commit adds lines to the `validate_csc_input` function that force
an in-place change of the input matrix to ensure that there are no
duplicates and the indices are sorted.

Add tests for non-canonical input (as of scipy v1.16.2), and ensure
`davis_example_chol` is returned as canonical format to isolate testing.