DOC: Update docs theme to Furo, per-method pages.
This commit uses the Furo theme to add a light/dark colorscheme button
and simplify the style. There was an issue with the PyData theme's
"search" feature on build, so use Furo for now.

This commit also adds the `sphinx.ext.autosummary` to create separate
pages for each function/method. It is much easier to read the
documentation with less scrolling.