DOC: Create UMFPACK docs.
This commit updates the docstrings of the umfpack module. It uses the
`autosummary` extension in Sphinx to generate individual pages for each
function and class in the umfpack module. It also adds an autosummary
template so that the docs for each class method are on their own html
pages, like the numpy and scipy docs.