TST/DOC: Test KLU refactorize with new structure.
KLU does *not* test for a structure change between the symbolic and
numeric factorization. It seems like a *new* factorization (i.e. no
existing numeric factor) will just work, but a *re*-factorization of an
existing numeric factor will fail silently, since it assumes the
structure is unchanged. Add a note in the documentation to this effect.