MNT: Change __cinit__ -> __init__. Remove CopySentinel.
This change leaves CholeskyFactor in an "unsafe" state after __cinit__
is run (or `CholeskyFactor.__new__(CholeskyFactor)`), since the `_cm`
and `_factor` pointers are NULL, but it would have to be very
intentional on the part of the user to create an object as such. There
is a warning in the docs.