MNT: Refactor copy functions to use a single dispatch.
This commit eliminates the need for a separate copy function for
klu_symbolic and klu_l_symbolic, as well as klu_numeric and
klu_l_numeric. It is a bit hacky, since we need an if-statement based on
the input types to generate unreachable code that will be pruned before
compilation, but it works and saves code duplication.